finish the hole projecttt

+++ ex01/Span.cpp
@@ -16,22 +16,22 @@ Span &Span::operator=(const Span &other)
 
 Span::~Span() {}
 
-void Span::addNumber(int num) {
+void Span::addNumber(long long num) {
     if (buf.size() >= Max_n)
         throw std::runtime_error("we cannot add more numbers");
     buf.push_back(num);
 }
 
-int Span::shortestSpan()
+long long Span::shortestSpan()
 {
     if (buf.size() < 2)
         throw std::runtime_error("Not enough numbers");
 
-    std::vector<int> tmp(buf);
+    std::vector<long long> tmp(buf);
     std::sort(tmp.begin(), tmp.end());
-    int shortest = tmp[1] - tmp[0];
+    long long shortest = tmp[1] - tmp[0];
     for (size_t i = 1; i < tmp.size() - 1; i++) {
-        int diff = tmp[i + 1] - tmp[i];
+        long long diff = tmp[i + 1] - tmp[i];
 
         if (diff < shortest)
             shortest = diff;
@@ -40,11 +40,11 @@ int Span::shortestSpan()
     return (shortest);
 }
 
-int Span::longestSpan() {
+long long Span::longestSpan() {
     if (buf.size() < 2)
         throw std::runtime_error("Not enough numbers");
 
-    int max = *std::max_element(buf.begin(), buf.end());
-    int min = *std::min_element(buf.begin(), buf.end());
+    long long max = *std::max_element(buf.begin(), buf.end());
+    long long min = *std::min_element(buf.begin(), buf.end());
     return max - min;
 }
